hello_cpp/requirements.txt
bottle==0.12.19
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
conan==1.34.0
deprecation==2.0.7
distro==1.5.0
fasteners==0.16
future==0.18.2
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
node-semver==0.6.1
packaging==20.9
patch-ng==1.17.4
pluginbase==1.0.0
Pygments==2.8.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
tqdm==4.58.0
urllib3==1.25.11
